+++ src/main/java/fr/alasdiablo/janoeo/JANOEO.java
@@ -2,14 +2,11 @@ package fr.alasdiablo.janoeo;
 
 import fr.alasdiablo.janoeo.holder.RegistryHolder;
 import fr.alasdiablo.janoeo.utils.*;
-import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 /**
  * Main class of the mods
@@ -21,11 +18,6 @@ public class JANOEO {
      * Mod setup
      */
     public static ModSetup setup = new ModSetup();
-
-    /**
-     * Mod logger (not use for the moment)
-     */
-    private static final Logger LOGGER = LogManager.getLogger();
 
     /**
      * Mod constructor